Create vpython_example.py
This is an example of a Sun - Earth (we can transform it to Earth - satellite) orbit visualization. We can use it for our project only in the end of the data analysis part, because as we can observe above it uses vectors of position and velocity for the ploting. So my personal opinion is that for the beggining of the project we can use the already existed matplotlib visualization and then try to implement this tool too for better 3d visualization. It needs the vpython library but it has produced some bugs too that needs fixing, like 


AttributeError: 'NoneType' object has no attribute 'kernel'  : when I tried to run it in Pycharm IDLE

<IPython.core.display.Javascript object> : when I tried to run it in Pycharm IDEL using a Jupyter Console

I managed to run the code properly in Jupyter Console seperately only.