Fixes issue 99 (#106)
* Fixes issue 99

Sets multi_rev to 0 in lamberts_kalman.py. This makes the output array
have the required shape.

Limits the cosine in state_kep.py to be between -1 and 1. Sometimes, due
to floating point errors, it can go out of limit.

Adds a new command line option to choose between metres and kilometres.

* Added cosine clipping to other cosines in the file.